Add files via upload
write an Android application that allows the nurses to keep track of various tests performed daily on patients in a hospital. Your application could be used by nurses in hospitals or testing centers. Use Room persistence library and MVVM pattern (as shown in class examples) to create and manipulate the application's database. (instead I used SQlite)